Organize tests and file structure
